+++ js/clarus.js
@@ -4,6 +4,7 @@ export { Tooltip } from "./tooltip.js";
 export { Popover } from "./popover.js";
 export { Modal } from "./modal.js";
 export { confirm } from "./confirm.js";
+export { Tag } from "./tag.js";
 export { Offcanvas } from "./offcanvas.js";
 export { Select } from "./select.js";
 export { Accordion } from "./accordion.js";